requirements.txt
crewai
python-dotenv
litellm
langchain_openai